give feedback on alias deletion when there is no matching alias

+++ commands/discord/ccmds/delalias.js
@@ -11,6 +11,9 @@ export const adminOnly = true;
 export async function run(client, message, args) {
     const twchannel = client.config.watchtimechannel;
     if (args.length == 0) return;
+    const alias = await client.clients.db.getAlias(twchannel, args[0]);
+    if (!alias) return message.reply({ content: `Ich kenne keinen Alias ${args[0]}.` });
+
     await client.clients.db.deleteAlias(twchannel, args[0]);
     message.reply({ content: `Der Alias ${args[0]} wurde entfernt.` });
     logger.log("command", `* Deleted alias ${args[0]}`);
