+++ commands/twitch/ccmds/delalias.js
@@ -18,6 +18,8 @@ export const alias = [];
 export async function run(client, target, context, msg, self, args) {
     const user = context["display-name"];
     if (args.length == 0) return;
+    const alias = await client.clients.db.getAlias(target, args[0]);
+    if (!alias) return client.say(target, `${user}, es gibt keinen Alias mit dem Namen ${args[0]}.`);;
     await client.clients.db.deleteAlias(target.replace(/#+/g, ""), args[0]);
     client.say(target, `${user}, der Alias ${args[0]} wurde entfernt.`);
     logger.log("command", `* Deleted alias ${args[0]}`);
